Refactor 4x4 keypad blocks to new module.
